+++ tests/nj_tests.cpp
@@ -2,7 +2,7 @@
 #include "../src/nj.cpp"
 
 TEST_CASE("nj with simple distance table", "[nj]"){
-    std::vector<float> d = {0.0,1.0,
+    std::vector<double> d = {0.0,1.0,
                             1.0,0.0};
     std::vector<std::string> l {"a", "b"};
     nj_t n(d,l);
@@ -12,7 +12,7 @@ TEST_CASE("nj with simple distance table", "[nj]"){
 }
 
 TEST_CASE("nj with larger distance table", "[nj]"){
-    std::vector<float> d = {0.0, 1.0, 1.0,
+    std::vector<double> d = {0.0, 1.0, 1.0,
                             1.0, 0.0, 1.0,
                             1.0, 1.0, 0.0};
     std::vector<std::string> l {"a", "b", "c"};
@@ -21,7 +21,7 @@ TEST_CASE("nj with larger distance table", "[nj]"){
 }
 
 TEST_CASE("nj with largest distance table", "[nj][regression]"){
-    std::vector<float> d = {0.0, 1.0, 2.5, 2.5,
+    std::vector<double> d = {0.0, 1.0, 2.5, 2.5,
                             1.0, 0.0, 2.5, 2.5,
                             2.5, 2.5, 0.0, 1.0,
                             2.5, 2.5, 1.0, 0.0};
@@ -31,12 +31,12 @@ TEST_CASE("nj with largest distance table", "[nj][regression]"){
 }
 
 TEST_CASE("nj and then setting weights", "[nj]"){
-    std::vector<float> d = {0.0,1.0,
+    std::vector<double> d = {0.0,1.0,
                             1.0,0.0};
     std::vector<std::string> l {"a", "b"};
     nj_t n(d,l);
     auto nj_tree = n.get_tree();
-    nj_tree.set_weights([](size_t) -> float {return 1.0;});
+    nj_tree.set_weights([](size_t) -> double {return 1.0;});
     nj_tree.sort();
     REQUIRE(nj_tree.to_string() == "(a:1.0,b:1.0);");
 }
